Move wallet viewmodels to subfolder
