+++ WalletWasabi.Fluent/ViewModels/Wallets/ClosedWatchOnlyWalletViewModel.cs
@@ -1,7 +1,7 @@
 using ReactiveUI;
 using WalletWasabi.Wallets;
 
-namespace WalletWasabi.Fluent.ViewModels
+namespace WalletWasabi.Fluent.ViewModels.Wallets
 {
 	public class ClosedWatchOnlyWalletViewModel : ClosedWalletViewModel
 	{
