+++ WalletWasabi.Fluent/ViewModels/Wallets/HardwareWalletViewModel.cs
@@ -2,7 +2,7 @@ using ReactiveUI;
 using WalletWasabi.Gui;
 using WalletWasabi.Wallets;
 
-namespace WalletWasabi.Fluent.ViewModels
+namespace WalletWasabi.Fluent.ViewModels.Wallets
 {
 	public class HardwareWalletViewModel : WalletViewModel
 	{
