+++ WalletWasabi.Fluent/ViewModels/Wallets/WalletViewModelBase.cs
@@ -6,7 +6,7 @@ using System.Reactive.Linq;
 using WalletWasabi.Helpers;
 using WalletWasabi.Wallets;
 
-namespace WalletWasabi.Fluent.ViewModels
+namespace WalletWasabi.Fluent.ViewModels.Wallets
 {
 	public abstract class WalletViewModelBase : NavBarItemViewModel, IComparable<WalletViewModelBase>, IDisposable
 	{
